requirements.txt
PyYAML>=6.0.0
openai>=1.30.0
requests>=2.31.0
Flask>=3.0.0
gunicorn>=22.0.0
pytest>=8.0.0
Pillow>=10.0.0
piexif>=1.1.3
iptcinfo3>=2.1.4
